+++ Demo/DemoUITests/CardUITests.swift
@@ -8,13 +8,13 @@ class CardUITests: XCTestCase {
     }
 
     func testApproveOrder() throws {
-        let app = launchApp(withArgs: ["-EnvironmentSandbox", "-DemoTypeCard"])
+        let app = launchApp(withArgs: ["-EnvironmentSandbox", "-DemoTypeCard", "-UIFrameworkUIKit"])
 
         app.button(named: "Create Order").tap()
 
         _ = app.staticText(containing: "Order ID:").waitForExistence()
 
-        app.textField(named: "Card Number").tap()
+        app.textField(named: "Card Number").tap() // this fails in SwiftUI, forcing test to run in UIKit
         app.typeText("5288775404117508")
 
         app.textField(named: "Expiration").tap()