+++ UnitTests/CardTests/ConfirmPaymentSourceRequest_Tests.swift
@@ -13,6 +13,7 @@ class ConfirmPaymentSourceRequest_Tests: XCTestCase {
             expirationYear: "2024",
             securityCode: "222"
         )
+        let bundleID = Bundle.main.bundleIdentifier ?? ""
         let cardRequest = CardRequest(orderID: mockOrderID, card: card)
 
         let confirmPaymentSourceRequest = try XCTUnwrap(
@@ -20,15 +21,32 @@ class ConfirmPaymentSourceRequest_Tests: XCTestCase {
         )
 
         let paymentSourceBody = try XCTUnwrap(confirmPaymentSourceRequest.body)
-        let paymentSourceBodyString = String(data: paymentSourceBody, encoding: .utf8)
-
-        // swiftlint:disable line_length
-        let expectedPaymentSourceBodyString = """
-        {"payment_source":{"card":{"number":"4032036247327321","security_code":"222","billing_address":null,"name":null,"attributes":null,"expiry":"2024-11"}}}
-        """
-        // swiftlint:enable line_length
-
-        XCTAssertEqual(paymentSourceBodyString, expectedPaymentSourceBodyString)
+        if let paymentSourceBodyString = String(data: paymentSourceBody, encoding: .utf8) {
+            let expectedPaymentSourceBodyString = """
+                {
+                    "application_context": {
+                        "return_url": "\(bundleID):\\/\\/card\\/success",
+                        "cancel_url": "\(bundleID):\\/\\/card\\/cancel"
+                    },
+                    "payment_source": {
+                        "card": {
+                            "number": "4032036247327321",
+                            "security_code": "222",
+                            "billing_address": null,
+                            "name": null,
+                            "attributes": {
+                                "verification": {
+                                    "method": "SCA_WHEN_REQUIRED"
+                                }
+                            },
+                            "expiry": "2024-11"
+                        }
+                    }
+                }
+                """.replacingOccurrences(of: " ", with: "").replacingOccurrences(of: "\n", with: "")
+            
+            XCTAssertEqual(paymentSourceBodyString, expectedPaymentSourceBodyString)
+        }
     }
 
     func testEncodingPaymentSource_withValidCardDictionary_expectsValidHTTPParams() throws {